+++ src/image/cache/google-cloud-storage/index.ts
@@ -42,7 +42,8 @@ export default class GoogleCloudStorageImageCache extends ImageCache {
   }
 
   public createKey (): string {
-    return createHash('md5').update(this.req.url).digest('hex')
+    const webpKey = this.config.imageable.action.supportWebp && this.req.headers.accept.includes('image/webp') ? 'webp' : ''
+    return createHash('md5').update(`${this.req.url}${webpKey}`).digest('hex')
   }
 
   public isValidFor (type) {