+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.12.4] - Unreleased
 
+### Added
+- Added webp support for local image action - @gibkigonzo (#409)
+
 ### Fixed 
 
  * add try catch to jwt.decode for groupId which is needed when you change the key for the JWT - @resubaka (#496)